Extract Safe Apps compatibility coordinator
